extract shared code
