+++ src/MQTTAnalyzer/mqtt/cocoamqtt/MQTT5ClientCocoaMQTT.swift
@@ -12,46 +12,29 @@ import Combine
 import Network
 
 class MQTT5ClientCocoaMQTT: MqttClient {
-	private let connectionStateQueue = DispatchQueue(label: "connection.state.lock.queue")
-	
 	let delgate = MQTT5Delegate()
-	
-	let sessionNum: Int
-	let model: TopicTree
-	var host: Host
 	var mqtt: CocoaMQTT5?
-	
+
+	let utils: ClientUtils
+	var connectionState: ConnectionState { utils.connectionState }
+	var host: Host { utils.host }
+
 	var connectionAlive: Bool {
 		self.mqtt != nil && connectionState.state == .connected
 	}
 	
-	var connectionState = ConnectionState()
-	
 	let messageSubject = MsgSubject<CocoaMQTT5Message>()
 		
 	init(host: Host, model: TopicTree) {
-		ConnectionState.sessionNum += 1
-
-		self.model = model
-		self.sessionNum = ConnectionState.sessionNum
-		self.host = host
+		utils = ClientUtils(host: host, model: model)
 	}
-	
-	func sanitizeBasePath(_ basePath: String) -> String {
-		if basePath.starts(with: "/") {
-			return basePath
-		}
-		else {
-			return "/\(basePath)"
-		}
-	}
-	
+		
 	func connect() {
-		initConnect()
+		utils.initConnect()
 		
 		let mqtt: CocoaMQTT5
 		if host.protocolMethod == .websocket {
-			let websocket = CocoaMQTTWebSocket(uri: sanitizeBasePath(self.host.basePath))
+			let websocket = CocoaMQTTWebSocket(uri: utils.sanitizeBasePath(self.host.basePath))
 			mqtt = CocoaMQTT5(clientID: host.computeClientID,
 								  host: host.hostname,
 								  port: host.port,
@@ -76,11 +59,11 @@ class MQTT5ClientCocoaMQTT: MqttClient {
 				try mqtt.sslSettings = createSSLSettings(host: host)
 			}
 			catch let error as CertificateError {
-				failConnection(reason: "\(error.rawValue)")
+				utils.failConnection(reason: "\(error.rawValue)")
 				return
 			}
 			catch {
-				failConnection(reason: "\(error)")
+				utils.failConnection(reason: "\(error)")
 				return
 			}
 		}
@@ -94,15 +77,15 @@ class MQTT5ClientCocoaMQTT: MqttClient {
 		mqtt.didConnectAck = self.didConnect
 		
 		if !mqtt.connect() {
-			failConnection(reason: "Connection to port \(host.port) failed")
+			utils.failConnection(reason: "Connection to port \(host.port) failed")
 			return
 		}
 
-		waitConnected()
+		utils.waitConnected()
 
 		self.mqtt = mqtt
 
-		let queue = DispatchQueue(label: "Message dispache queue")
+		let queue = DispatchQueue(label: "Message Dispatch queue")
 		messageSubject.cancellable = messageSubject.subject.eraseToAnyPublisher()
 			.collect(.byTime(queue, 0.5))
 			.receive(on: DispatchQueue.main)
@@ -110,73 +93,9 @@ class MQTT5ClientCocoaMQTT: MqttClient {
 				self.onMessageInMain(messages: $0)
 			})
 	}
-
-	// MARK: Should be shared
-	func waitConnected() {
-
-		let group = DispatchGroup()
-		group.enter()
-
-		DispatchQueue.global().async {
-			var i = 10
-			
-			var connecting = true
-			
-			while connecting && i > 0 {
-				print("CONNECTION: waiting... \(self.sessionNum) \(i) \(self.host.hostname)")
-				sleep(1)
-				
-				i-=1
-				
-				self.connectionStateQueue.sync {
-					connecting = self.connectionState.state == .connecting
-				}
-			}
-			group.leave()
-		}
-
-		group.notify(queue: .main) {
-			if let errorMessage = self.connectionState.message {
-				self.setDisconnected()
-				self.host.connectionMessage = errorMessage
-				return
-			}
-
-			if self.host.state != .connected {
-				self.setDisconnected()
-
-				self.setConnectionMessage(message: "Connection timeout")
-			}
-		}
-	}
 	
-	// MARK: Should be shared
-	func setConnectionMessage(message: String) {
-		DispatchQueue.global(qos: .userInitiated).async {
-			DispatchQueue.main.async {
-				self.host.connectionMessage = message
-			}
-		}
-	}
-	
-	// MARK: Should be shared
-	func initConnect() {
-		print("CONNECTION: connect \(sessionNum) \(host.hostname)")
-		host.connectionMessage = nil
-		host.state = .connecting
-		connectionState.state = .connecting
-		connectionState.message = nil
-		model.messageLimitExceeded = false
-		model.topicLimitExceeded = false
-	}
-		
 	func disconnect() {
-		print("CONNECTION: disconnect \(sessionNum) \(host.hostname)")
-
 		messageSubject.cancel()
-		self.connectionStateQueue.async {
-			self.connectionState.state = .disconnected
-		}
 		
 		if let mqtt = self.mqtt {
 			DispatchQueue.global(qos: .background).async {
@@ -184,8 +103,6 @@ class MQTT5ClientCocoaMQTT: MqttClient {
 				mqtt.disconnect()
 
 				DispatchQueue.main.async {
-					print("CONNECTION: disconnected \(self.sessionNum) \(self.host.hostname)")
-					
 					self.setDisconnected()
 				}
 			}
@@ -197,144 +114,46 @@ class MQTT5ClientCocoaMQTT: MqttClient {
 		let message = CocoaMQTT5Message(
 			topic: message.topic.nameQualified,
 			string: message.payload.dataString,
-			qos: convertQOS(qos: message.metadata.qos),
+			qos: utils.convertQOS(qos: message.metadata.qos),
 			retained: message.metadata.retain
 		)
 		message.contentType = "application/json"
 		mqtt?.publish(message, properties: properties)
 	}
-
-	func convertQOS(qos: Int32) -> CocoaMQTTQoS {
-		switch qos {
-		case 1:
-			return CocoaMQTTQoS.qos1
-		case 2:
-			return CocoaMQTTQoS.qos2
-		default:
-			return CocoaMQTTQoS.qos0
-		}
-	}
 	
 	func setDisconnected() {
-		self.connectionStateQueue.async {
-			self.connectionState.state = .disconnected
-		}
-
-		DispatchQueue.main.async {
-			self.host.state = .disconnected
-		}
+		utils.setDisconnected()
 		mqtt = nil
-	}
-
-	// MARK: Should be shared
-	func onMessageInMain(messages: [CocoaMQTT5Message]) {
-		if host.pause {
-			return
-		}
-		
-		//		model.limitMessagesPerBatch = host.limitMessagesBatch
-		//		model.limitTopics = host.limitTopic
-		if messages.count > host.limitMessagesBatch {
-			// Limit exceeded
-			self.model.messageLimitExceeded = true
-			return
-		}
-		
-		for message in messages {
-			if host.limitTopic > 0 && self.model.totalTopicCounter >= host.limitTopic {
-				// Limit exceeded
-				self.model.topicLimitExceeded = true
-			}
-			
-			_ = self.model.addMessage(
-				metadata: MsgMetadata(qos: Int32(message.qos.rawValue), retain: message.retained),
-				payload: MsgPayload(data: message.payload),
-				to: message.topic
-			)
-		}
 	}
 	
 	// MARK: Should be shared
 	func subscribeToTopic(_ host: Host) {
 		host.subscriptions.forEach {
-			mqtt?.subscribe($0.topic, qos: convertQOS(qos: Int32($0.qos)))
+			mqtt?.subscribe($0.topic, qos: utils.convertQOS(qos: Int32($0.qos)))
 		}
 	}
 
 	func didDisconnect(_ mqtt: CocoaMQTT5, withError err: Error?) {
-		print("CONNECTION: onDisconnect \(sessionNum) \(host.hostname)")
-
-		if err != nil {
-			let messgae = MQTTClientCocoaMQTT.extractErrorMessage(error: err!)
-			
-			self.connectionStateQueue.async {
-				self.connectionState.message = messgae
-			}
-			DispatchQueue.main.async {
-				self.host.usernameNonpersistent = nil
-				self.host.passwordNonpersistent = nil
-				self.host.connectionMessage = self.connectionState.message
-			}
-		}
-		
-		self.setDisconnected()
-
-		DispatchQueue.main.async {
-			self.host.pause = false
-			self.host.state = .disconnected
-		}
+		utils.didDisconnect(withError: err)
 	}
 	
 	func didConnect(_ mqtt: CocoaMQTT5, reasonCode: CocoaMQTTCONNACKReasonCode, didConnectAck ack: MqttDecodeConnAck) {
 		
 		switch reasonCode {
 		case .success:
-			connectedSuccess(didConnectAck: ack)
+			utils.connectedSuccess()
+			subscribeToTopic(host)
 		case .badUsernameOrPassword:
-			clearAuth()
-			failConnection(reason: "Bad username/password")
+			utils.clearAuth()
+			utils.failConnection(reason: "Bad username/password")
 		case .notAuthorized:
-			clearAuth()
-			failConnection(reason: "Not authorized")
+			utils.clearAuth()
+			utils.failConnection(reason: "Not authorized")
 		default:
-			failConnection(reason: String(describing: reasonCode))
+			utils.failConnection(reason: String(describing: reasonCode))
 		}
 	}
-	
-	func connectedSuccess(didConnectAck ack: MqttDecodeConnAck) {
-		print("CONNECTION: onConnect \(sessionNum) \(host.hostname)")
-		self.connectionStateQueue.async {
-			self.connectionState.state = .connected
-		}
-
-		NSLog("Connected. Return Code is \(ack.description)")
-		DispatchQueue.main.async {
-			self.host.state = .connected
-		}
 		
-		subscribeToTopic(host)
-	}
-	
-	func clearAuth() {
-		self.host.usernameNonpersistent = nil
-		self.host.passwordNonpersistent = nil
-	}
-	
-	func failConnection(reason: String) {
-		NSLog("Connection failed: " + reason)
-		self.connectionStateQueue.async {
-			self.connectionState.message = reason
-		}
-
-		self.setDisconnected()
-
-		DispatchQueue.main.async {
-			self.host.connectionMessage = reason
-			self.host.pause = false
-			self.host.state = .disconnected
-		}
-	}
-	
 	func didReceiveMessage(client: CocoaMQTT5, message: CocoaMQTT5Message, qos: UInt16, decode: MqttDecodePublish) {
 		if !host.pause {
 			messageSubject.send(message)
