+++ src/MQTTAnalyzerTests/MqttClientCocoaMQTTTests.swift
@@ -13,22 +13,22 @@ import Network
 class MqttClientCocoaMQTTTests: XCTestCase {
 	 
 	func testInvalidHostname() throws {
-		let msg = MqttClientCocoaMQTT.extractErrorMessage(error: NSError(domain: "", code: 8))
+		let msg = ClientUtils.extractErrorMessage(error: NSError(domain: "", code: 8))
 		XCTAssertEqual("Invalid hostname.\nThe operation couldn’t be completed. ( error 8.)", msg)
 	}
 	
     func testConnectionRefused() throws {
-		let msg = MqttClientCocoaMQTT.extractErrorMessage(error: NWError.posix(POSIXErrorCode.ECONNREFUSED))
+		let msg = ClientUtils.extractErrorMessage(error: NWError.posix(POSIXErrorCode.ECONNREFUSED))
 		XCTAssertTrue(msg.contains("Connection refused"), msg)
 	}
 	
 	func testConnectionTLS() throws {
-		let msg = MqttClientCocoaMQTT.extractErrorMessage(error: NWError.tls(-9407))
+		let msg = ClientUtils.extractErrorMessage(error: NWError.tls(-9407))
 		XCTAssertEqual("Network.NWError: -9407: Optional(OSStatus -9407)", msg)
 	}
 	
 	func testConnectionTLSBadCertificate() throws {
-		let msg = MqttClientCocoaMQTT.extractErrorMessage(error: NWError.tls(-9808))
+		let msg = ClientUtils.extractErrorMessage(error: NWError.tls(-9808))
 		XCTAssertEqual("Bad certificate format, check all properties, like SAN, ... (-9808)", msg)
 	}
